requirements-darwin.txt
foohid
pyobjc
pyobjc-framework-Quartz
